gcp_function/requirements.txt
functions-framework==3.4.0
tensorflow==2.15.0
Pillow>=9.0
numpy
